+++ packages/react/src/hooks/useFocusToggle.ts
@@ -1,4 +1,9 @@
-import { setupFocusToggle, isTrackReferencePinned } from '@livekit/components-core';
+import type { TrackReferenceOrPlaceholder } from '@livekit/components-core';
+import {
+  setupFocusToggle,
+  isTrackReferencePinned,
+  isTrackReference,
+} from '@livekit/components-core';
 import type { Track, Participant } from 'livekit-client';
 import { useEnsureParticipant, useMaybeLayoutContext } from '../context';
 import { mergeProps } from '../mergeProps';
@@ -6,28 +11,50 @@ import * as React from 'react';
 
 /** @public */
 export interface UseFocusToggleProps {
-  props: React.ButtonHTMLAttributes<HTMLButtonElement>;
-  trackSource: Track.Source;
+  trackRef?: TrackReferenceOrPlaceholder;
+  /** @deprecated This parameter will be removed in a future version use `trackRef` instead. */
+  trackSource?: Track.Source;
+  /** @deprecated This parameter will be removed in a future version use `trackRef` instead. */
   participant?: Participant;
+  props: React.ButtonHTMLAttributes<HTMLButtonElement>;
 }
 
-/** @public */
-export function useFocusToggle({ trackSource, participant, props }: UseFocusToggleProps) {
+/**
+ * The `useFocusToggle` hook is used to implement the `FocusToggle` or your custom implementation of it.
+ * The `TrackReferenceOrPlaceholder` is used to register a onClick handler and to identify the track to focus on.
+ *
+ * @example
+ * ```tsx
+ * const { mergedProps, inFocus } = useFocusToggle({ trackRef, props: yourButtonProps });
+ * return <button {...mergedProps}>{inFocus ? 'Unfocus' : 'Focus'}</button>;
+ * ```
+ * @public
+ */
+export function useFocusToggle({ trackRef, trackSource, participant, props }: UseFocusToggleProps) {
   const p = useEnsureParticipant(participant);
+  if (!trackRef && !trackSource) {
+    throw new Error('trackRef or trackSource must be defined.');
+  }
   const layoutContext = useMaybeLayoutContext();
   const { className } = React.useMemo(() => setupFocusToggle(), []);
 
   const inFocus: boolean = React.useMemo(() => {
-    const track = p.getTrack(trackSource);
-    if (layoutContext?.pin.state && track) {
-      return isTrackReferencePinned(
-        { participant: p, source: trackSource, publication: track },
-        layoutContext.pin.state,
-      );
+    if (trackRef) {
+      return isTrackReferencePinned(trackRef, layoutContext?.pin.state);
+    } else if (trackSource) {
+      const track = p.getTrack(trackSource);
+      if (layoutContext?.pin.state && track) {
+        return isTrackReferencePinned(
+          { participant: p, source: trackSource, publication: track },
+          layoutContext.pin.state,
+        );
+      } else {
+        return false;
+      }
     } else {
-      return false;
+      throw new Error('trackRef or trackSource and participant must be defined.');
     }
-  }, [p, trackSource, layoutContext]);
+  }, [trackRef, layoutContext?.pin.state, p, trackSource]);
 
   const mergedProps = React.useMemo(
     () =>
@@ -38,26 +65,39 @@ export function useFocusToggle({ trackSource, participant, props }: UseFocusTogg
           props.onClick?.(event);
 
           // Set or clear focus based on current focus state.
-          const track = p.getTrack(trackSource);
-          if (layoutContext?.pin.dispatch && track) {
+          if (trackRef && isTrackReference(trackRef)) {
             if (inFocus) {
-              layoutContext.pin.dispatch({
+              layoutContext?.pin.dispatch?.({
                 msg: 'clear_pin',
               });
             } else {
-              layoutContext.pin.dispatch({
+              layoutContext?.pin.dispatch?.({
                 msg: 'set_pin',
-                trackReference: {
-                  participant: p,
-                  publication: track,
-                  source: track.source,
-                },
+                trackReference: trackRef,
               });
+            }
+          } else if (trackSource) {
+            const track = p.getTrack(trackSource);
+            if (layoutContext?.pin.dispatch && track) {
+              if (inFocus) {
+                layoutContext.pin.dispatch({
+                  msg: 'clear_pin',
+                });
+              } else {
+                layoutContext.pin.dispatch({
+                  msg: 'set_pin',
+                  trackReference: {
+                    participant: p,
+                    publication: track,
+                    source: track.source,
+                  },
+                });
+              }
             }
           }
         },
       }),
-    [props, className, p, trackSource, inFocus, layoutContext],
+    [props, className, trackRef, trackSource, inFocus, layoutContext?.pin, p],
   );
 
   return { mergedProps, inFocus };