+++ packages/react/src/prefabs/ControlBar.tsx
@@ -1,6 +1,7 @@
 import { Track } from 'livekit-client';
 import * as React from 'react';
 import { MediaDeviceMenu } from './MediaDeviceMenu';
+import { HostEndMeetingMenu } from './HostEndMeetingMenu';
 import { DisconnectButton } from '../components/controls/DisconnectButton';
 import { TrackToggle } from '../components/controls/TrackToggle';
 import { StartAudio } from '../components/controls/StartAudio';
@@ -12,7 +13,7 @@ import SvgInviteIcon from '../assets/icons/InviteIcon';
 import SvgUserIcon from '../assets/icons/UsersIcon';
 import { useLocalParticipantPermissions } from '../hooks';
 import { useMediaQuery } from '../hooks/internal';
-import { useMaybeLayoutContext } from '../context';
+import { useLayoutContext, useMaybeLayoutContext } from '../context';
 import { supportsScreenSharing } from '@livekit/components-core';
 import { mergeProps } from '../utils';
 
@@ -26,6 +27,7 @@ export type ControlBarControls = {
   sharelink?: boolean;
   users?: boolean;
   leaveButton?: string;
+  endForAll?: string | false;
 };
 
 /** @public */
@@ -37,8 +39,8 @@ export interface ControlBarProps extends React.HTMLAttributes<HTMLDivElement> {
 };
 
 /**
- * The ControlBar prefab component gives the user the basic user interface
- * to control their media devices and leave the room.
+ * The `ControlBar` prefab gives the user the basic user interface to control their
+ * media devices (camera, microphone and screen share), open the `Chat` and leave the room.
  *
  * @remarks
  * This component is build with other LiveKit components like `TrackToggle`,
@@ -60,10 +62,10 @@ export function ControlBar({
   ...props
 }: ControlBarProps) {
   const layoutContext = useMaybeLayoutContext();
-
   const [isChatOpen, setIsChatOpen] = React.useState(false);
   const [isShareLinkOpen, setIsShareLinkOpen] = React.useState(false);
   const [isUserOpen, setIsUserOpen] = React.useState(false);
+  const { state } = useLayoutContext().widget;
 
   React.useEffect(() => {
     if (layoutContext?.widget.state?.showChat == 'show_chat') {
@@ -120,6 +122,34 @@ export function ControlBar({
   };
 
   const htmlProps = mergeProps({ className: 'lk-control-bar' }, props);
+  console.log(`Share scree tracks ${screenShareTracks}`);
+  React.useEffect(() => {
+    // Get all button elements with the "data-lk-source" attribute
+    const buttons = document.querySelectorAll('[data-lk-source]');
+    if (!isScreenShareEnabled && screenShareTracks !== 0) {
+      // Loop through each button and check its "data-lk-source" attribute
+      buttons.forEach(button => {
+        const source = button.getAttribute('data-lk-source');
+
+        // Check if the "data-lk-source" attribute value is "screen_share"
+        if (source === 'screen_share') {
+          // Disable the button
+          (button as HTMLButtonElement).disabled = true;
+        }
+      });
+    } else {
+      // Loop through each button and check its "data-lk-source" attribute
+      buttons.forEach(button => {
+        const source = button.getAttribute('data-lk-source');
+
+        // Check if the "data-lk-source" attribute value is "screen_share"
+        if (source === 'screen_share') {
+          // Disable the button
+          (button as HTMLButtonElement).disabled = false;
+        }
+      });
+    }
+  }, [screenShareTracks, isScreenShareEnabled]);
 
   React.useEffect(() => {
     // Get all button elements with the "data-lk-source" attribute
@@ -161,6 +191,7 @@ export function ControlBar({
           </div>
         </div>
       )}
+
       {visibleControls.camera && (
         <div className="lk-button-group">
           <TrackToggle source={Track.Source.Camera} showIcon={showIcon}>
@@ -171,6 +202,7 @@ export function ControlBar({
           </div>
         </div>
       )}
+
       {visibleControls.screenShare && browserSupportsScreenSharing && (
         <TrackToggle
           source={Track.Source.ScreenShare}
@@ -193,6 +225,12 @@ export function ControlBar({
         <ChatToggle>
           {showIcon && <ChatIcon />}
           {showText && 'Chat'}
+          {state && state.unreadMessages !== 0 && (
+            <span className="waiting-count">
+              {state.unreadMessages < 10
+                ? (state.unreadMessages.toFixed(0))
+                : ('9+')}
+            </span>)}
         </ChatToggle>
       )}
       {visibleControls.sharelink && (
@@ -208,13 +246,30 @@ export function ControlBar({
           {waitingRoomCount !== 0 && <span className="waiting-count">{waitingRoomCount}</span>}
         </UserToggle>
       )}
-      {visibleControls.leave && (
+      {visibleControls.endForAll ? (
+        <div className="tl-leave lk-button-group">
+          <button className="lk-disconnect-button">
+            {showIcon && <LeaveIcon />}
+            {showText && "Leave"}
+          </button>
+
+          <div className="tl-leave-btn lk-button-group-menu">
+            <HostEndMeetingMenu
+              leave={visibleControls.leave}
+              leaveButton={visibleControls.leaveButton}
+              endForAll={visibleControls.endForAll}
+              showIcon={showIcon}
+              showText={showText}
+            />
+          </div>
+        </div>
+      ) : (
         <DisconnectButton>
           {showIcon && <LeaveIcon />}
           {showText && visibleControls.leaveButton}
         </DisconnectButton>
       )}
       <StartAudio label="Start Audio" />
-    </div>
+    </div >
   );
 }