+++ packages/react/src/prefabs/Users.tsx
@@ -8,12 +8,12 @@ import type { LocalUserChoices } from './PreJoin';
 import { ToggleSwitch } from '../components/ToggleSwitch';
 import SvgApproveIcon from '../assets/icons/ApproveIcon';
 import SvgRejectIcon from '../assets/icons/RejectIcon';
-import { getToken } from './ShareLink';
+import { getDomainIdentifier, getToken } from './ShareLink';
 
 /** @public */
 export interface UserProps extends React.HTMLAttributes<HTMLDivElement> {
   onWaitingRoomChange: (state: number) => void;
-  setWaiting: (state: string) => void;
+  // setWaiting: (state: string) => void;
 }
 
 export type UserDataProps = {
@@ -42,6 +42,32 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
   /**
    * Get list of users in waiting room
    */
+  async function getWaitingRoomState() {
+    const postData = {
+      method: 'POST',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        meeting_id: room.name,
+        domain: getDomainIdentifier()
+      }),
+    };
+    fetch(`/api/get-waitingroom-state`, postData).then(async (res) => {
+      if (res.ok) {
+        const body = await res.json();
+        console.log(body);
+
+        setToggleWaiting(body.waiting_room);
+      } else {
+        throw Error('Error fetching server url, check server logs');
+      }
+    });
+  }
+  /**
+   * Get list of users in waiting room
+   */
   async function usersList() {
     const postData = {
       method: 'POST',
@@ -52,6 +78,7 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
       body: JSON.stringify({
         meeting_id: room.name,
         token: getToken(),
+        domain: getDomainIdentifier()
       }),
     };
     fetch(`/api/get-waiting-room-users`, postData).then(async (res) => {
@@ -77,6 +104,10 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
       // setWaiting(`Users are in waiting room`)
     }, 2000)
     return () => clearInterval(interval);
+  }, []);
+
+  React.useEffect(() => {
+    getWaitingRoomState();
   }, []);
 
   React.useEffect(() => {
@@ -108,6 +139,7 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
         identity: identity,
         type: type,
         token: getToken(),
+        domain: getDomainIdentifier()
       }),
     };
 
@@ -136,7 +168,8 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
       },
       body: JSON.stringify({
         meeting_id: room.name,
-        token: getToken()
+        token: getToken(),
+        domain: getDomainIdentifier()
       }),
     };
 
@@ -164,7 +197,8 @@ export function Users({ onWaitingRoomChange, ...props }: UserProps) {
       body: JSON.stringify({
         room: room.name,
         waiting_room: checked,
-        token: getToken()
+        token: getToken(),
+        domain: getDomainIdentifier()
       }),
     };