+++ packages/react/src/prefabs/VideoConference.tsx
@@ -1,43 +1,50 @@
-import * as React from 'react';
-import { LayoutContextProvider } from '../components/layout/LayoutContextProvider';
-import { RoomAudioRenderer } from '../components/RoomAudioRenderer';
-import { ControlBar } from './ControlBar';
-import { FocusLayout, FocusLayoutContainer } from '../components/layout/FocusLayout';
-import { GridLayout } from '../components/layout/GridLayout';
-import type { WidgetState } from '@livekit/components-core';
-import { ShareLink } from './ShareLink';
-import { Users } from './Users';
-import { isEqualTrackRef, isTrackReference, log, isWeb } from '@livekit/components-core';
-// import { Chat } from './Chat';
-import { ConnectionStateToast } from '../components/Toast';
-// import type { MessageFormatter } from '../components/ChatEntry';
+import type {
+  MessageDecoder,
+  MessageEncoder,
+  TrackReferenceOrPlaceholder,
+  WidgetState,
+} from '@livekit/components-core';
+import { isEqualTrackRef, isTrackReference, isWeb, log, setupParticipantName } from '@livekit/components-core';
 import { RoomEvent, Track } from 'livekit-client';
-import { useTracks } from '../hooks/useTracks';
-import { usePinnedTracks } from '../hooks/usePinnedTracks';
-import { CarouselLayout } from '../components/layout/CarouselLayout';
-import { useCreateLayoutContext } from '../context/layout-context';
-import { ParticipantTile } from '../components';
-import { Toast } from '../components';
-import { UserToggle } from '../components/controls/UserToggle';
-import type { TrackReferenceOrPlaceholder } from '@livekit/components-core';
+import * as React from 'react';
+import type { MessageFormatter } from '../components';
+import {
+  CarouselLayout,
+  ConnectionStateToast,
+  FocusLayout,
+  FocusLayoutContainer,
+  GridLayout,
+  LayoutContextProvider,
+  ParticipantTile,
+  RoomAudioRenderer,
+  formatChatMessageLinks,
+} from '../components';
+import { useCreateLayoutContext, useEnsureParticipant } from '../context';
+import { useLocalParticipant, usePinnedTracks, useTracks } from '../hooks';
+import { Chat } from './Chat';
+import { ControlBar } from './ControlBar';
+import { Users } from './Users';
+import { ShareLink } from './ShareLink';
+import { useObservableState } from '../hooks/internal';
 
 /**
  * @public
  */
 export interface VideoConferenceProps extends React.HTMLAttributes<HTMLDivElement> {
-  // chatMessageFormatter?: MessageFormatter;
-  showShareButton: boolean;
-  showParticipantButton: boolean;
-  leaveButton: string;
+  chatMessageFormatter?: MessageFormatter;
+  chatMessageEncoder?: MessageEncoder;
+  chatMessageDecoder?: MessageDecoder;
 }
 
 /**
- * This component is the default setup of a classic LiveKit video conferencing app.
- * It provides functionality like switching between participant grid view and focus view.
+ * The `VideoConference` ready-made component is your drop-in solution for a classic video conferencing application.
+ * It provides functionality such as focusing on one participant, grid view with pagination to handle large numbers
+ * of participants, basic non-persistent chat, screen sharing, and more.
  *
  * @remarks
  * The component is implemented with other LiveKit components like `FocusContextProvider`,
  * `GridLayout`, `ControlBar`, `FocusLayoutContainer` and `FocusLayout`.
+ * You can use this components as a starting point for your own custom video conferencing application.
  *
  * @example
  * ```tsx
@@ -48,18 +55,38 @@ export interface VideoConferenceProps extends React.HTMLAttributes<HTMLDivElemen
  * @public
  */
 export function VideoConference({
-  showShareButton,
-  showParticipantButton,
-  leaveButton,
+  chatMessageFormatter,
+  chatMessageDecoder,
+  chatMessageEncoder,
   ...props
 }: VideoConferenceProps) {
   const [widgetState, setWidgetState] = React.useState<WidgetState>({
     showChat: null,
     unreadMessages: 0,
   });
-  const lastAutoFocusedScreenShareTrack = React.useRef<TrackReferenceOrPlaceholder | null>(null);
 
-  const [waiting, setWaiting] = React.useState<string | null>(null); // Used to show toast message
+  const lastAutoFocusedScreenShareTrack = React.useRef<TrackReferenceOrPlaceholder | null>(null);
+  const { localParticipant } = useLocalParticipant();
+  const p = useEnsureParticipant(localParticipant);
+
+  const { infoObserver } = React.useMemo(() => {
+    return setupParticipantName(p);
+  }, [p]);
+
+  const { metadata } = useObservableState(infoObserver, {
+    name: p.name,
+    identity: p.identity,
+    metadata: p.metadata,
+  });
+
+  const [showShareButton, setShowShareButton] = React.useState<boolean>(false);
+  const [showParticipantButton, setShowParticipantButton] = React.useState<boolean>(false);
+  const [leaveButton, setLeaveButton] = React.useState<string>("Leave");
+  const [endForAll, setEndForAll] = React.useState<string | false>(false);
+
+  const meta = metadata ? JSON.parse(metadata) : {};
+
+  // const [waiting, setWaiting] = React.useState<string | null>(null); // Used to show toast message
   const [waitingRoomCount, setWaitingRoomCount] = React.useState<number>(0);
 
   const tracks = useTracks(
@@ -72,6 +99,8 @@ export function VideoConference({
 
   const widgetUpdate = (state: WidgetState) => {
     log.debug('updating widget state', state);
+    console.log(state);
+
     setWidgetState(state);
   };
 
@@ -80,11 +109,11 @@ export function VideoConference({
     setWaitingRoomCount(count);
   };
 
-  const setWaitingMessage = (message: string) => {
-    if (showParticipantButton) {
-      setWaiting(message);
-    }
-  };
+  // const setWaitingMessage = (message: string) => {
+  //   if (showParticipantButton) {
+  //     setWaiting(message);
+  //   }
+  // };
 
   const layoutContext = useCreateLayoutContext();
   const screenShareTracks = tracks
@@ -94,18 +123,40 @@ export function VideoConference({
   const focusTrack = usePinnedTracks(layoutContext)?.[0];
   const carouselTracks = tracks.filter((track) => !isEqualTrackRef(track, focusTrack));
 
+  // React.useEffect(() => {
+  //   if (waiting) {
+  //     // Remove toast message after 2 second
+  //     setTimeout(() => {
+  //       setWaiting(null);
+  //     }, 3000);
+  //   }
+  // }, [waiting]);
+
   React.useEffect(() => {
-    if (waiting) {
-      // Remove toast message after 2 second
-      setTimeout(() => {
-        setWaiting(null);
-      }, 3000);
+    if (meta && meta.host) {
+      setShowShareButton(true);
+      setShowParticipantButton(true);
+      setLeaveButton("Leave Meeting");
+      setEndForAll("End Meeting for All");
     }
-  }, [waiting]);
+  }, [meta]);
+
+  React.useEffect(() => {
+    const pmeta = p.metadata ? JSON.parse(p.metadata) : {};
+    if (pmeta && pmeta.host) {
+      setShowShareButton(true);
+      setShowParticipantButton(true);
+      setLeaveButton("Leave Meeting");
+      setEndForAll("End Meeting for All");
+    }
+  }, [p]);
 
   React.useEffect(() => {
     // If screen share tracks are published, and no pin is set explicitly, auto set the screen share.
-    if (screenShareTracks.length > 0 && lastAutoFocusedScreenShareTrack.current === null) {
+    if (
+      screenShareTracks.some((track) => track.publication.isSubscribed) &&
+      lastAutoFocusedScreenShareTrack.current === null
+    ) {
       log.debug('Auto set screen share focus:', { newScreenShareTrack: screenShareTracks[0] });
       layoutContext.pin.dispatch?.({ msg: 'set_pin', trackReference: screenShareTracks[0] });
       lastAutoFocusedScreenShareTrack.current = screenShareTracks[0];
@@ -122,7 +173,9 @@ export function VideoConference({
       lastAutoFocusedScreenShareTrack.current = null;
     }
   }, [
-    screenShareTracks.map((ref) => ref.publication.trackSid).join(),
+    screenShareTracks
+      .map((ref) => `${ref.publication.trackSid}_${ref.publication.isSubscribed}`)
+      .join(),
     focusTrack?.publication?.trackSid,
   ]);
 
@@ -147,30 +200,30 @@ export function VideoConference({
                   <CarouselLayout tracks={carouselTracks}>
                     <ParticipantTile />
                   </CarouselLayout>
-                  {focusTrack && <FocusLayout track={focusTrack} />}
+                  {focusTrack && <FocusLayout trackRef={focusTrack} />}
                 </FocusLayoutContainer>
               </div>
             )}
             <ControlBar
               controls={{
-                chat: false,
+                chat: true,
                 sharelink: showShareButton,
                 users: showParticipantButton,
                 leaveButton: leaveButton,
+                endForAll: endForAll,
               }}
               waitingRoomCount={waitingRoomCount}
               screenShareTracks={screenShareTracks.length}
             />
-          </div >
-
-
+          </div>
 
           {
             showShareButton ?
               (
                 <ShareLink style={{
                   display: widgetState.showChat == 'show_invite' ? 'flex' : 'none'
-                }} />
+                }
+                } />
               ) : (
                 <></>
               )
@@ -181,21 +234,18 @@ export function VideoConference({
               <Users
                 style={{ display: widgetState.showChat == 'show_users' ? 'flex' : 'none' }}
                 onWaitingRoomChange={updateCount}
-                setWaiting={setWaitingMessage}
               />
+              // setWaiting={setWaitingMessage}
             ) : (<></>)
           }
 
-          {
-            waiting ? (
-              <Toast className="lk-toast-connection-state">
-                <UserToggle>{waiting}</UserToggle>
-              </Toast>
-            ) : (
-              <></>
-            )
-          }
-        </LayoutContextProvider>
+          <Chat
+            style={{ display: widgetState.showChat == 'show_chat' ? 'flex' : 'none' }}
+            messageFormatter={formatChatMessageLinks}
+            messageEncoder={chatMessageEncoder}
+            messageDecoder={chatMessageDecoder}
+          />
+        </LayoutContextProvider >
       )
       }
       <RoomAudioRenderer />